# enhance: Optimize the performance of stats task