… internal/flushcommon/io/mock_binlogio.go
… internal/flushcommon/pipeline/mock_fgmanager.go
… internal/flushcommon/syncmgr/mock_meta_writer.go
… internal/flushcommon/syncmgr/mock_serializer.go
… internal/flushcommon/syncmgr/mock_task.go
… internal/flushcommon/writebuffer/mock_manager.go
… internal/flushcommon/writebuffer/mock_write_buffer.go