1. Increase the writer's `batchSize` to avoid multiple serialization operations.
2. Perform asynchronous upload of binlog files to prevent blocking the data processing flow.
3. Reduce multiple calls to `writer.Flush()`.

issue: https://github.com/milvus-io/milvus/issues/37373