@bigsheeper 2024-11-01
These optimizations reduced the stats task time from 15s to 7s (1 million 128-dim end-to-end test).
Before:
```
[2024/11/01 19:48:11.424 +08:00] [INFO] [indexnode/task_stats.go:301] ["sort segment end"] [clusterID=by-dev] [taskID=453630131593571107] [collectionID=453630131591970875] [partitionID=453630131591970876] [segmentID=453630131592170892] [subTaskType=Sort] ["target segmentID"=453630131593571106] ["old rows"=1000000] ["valid rows"=1000000] ["binlog batch count"=8] ["download elapse"=3.001232668s] ["upload binlogs elapse"=4.072131613s] ["sort elapse"=417.540613ms] ["serWrite elapse"=671.678849ms] ["write elapse"=7.321798417s] ["total1 elapse"=15.525105081s] ["total elapse"=15.524792759s]
```
After:
```
[2024/11/01 20:06:34.003 +08:00] [INFO] [indexnode/task_stats.go:296] ["sort segment end"] [clusterID=by-dev] [taskID=453630403454556934] [collectionID=453630403453156710] [partitionID=453630403453156711] [segmentID=453630403453156728] [subTaskType=Sort] ["target segmentID"=453630403454556933] ["old rows"=1000000] ["valid rows"=1000000] ["binlog batch count"=8] ["download elapse"=2.640881298s] ["upload binlogs elapse"=0s] ["sort elapse"=355.944256ms] ["serWrite elapse"=546.764809ms] ["write elapse"=3.359702497s] ["total elapse"=7.058947854s]
```